requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
attrs==24.1.0
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
py-cord==2.6.0
python-dotenv==1.0.1
toml==0.10.2
yarl==1.9.4
